refactor(mcp): extract proxy tool execute handler into proxy-tool.ts
